Low-level API

The Ginkgo.API submodule provides a low-level interface which closely matches the Ginkgo C API. While these functions are not intended for general usage, they are useful for calling Ginkgo routines not yet available in Ginkgo.jl main interface, and is the basis for the high-level wrappers. For illustrative purpose, we use a example api.jl here, yet to be replaced.

Ginkgo.API.get_version — Method

get_version()

Get the version of the package

Ginkgo.API.integrate — Method

integrate(a, b, bins, _function)

Integrate over a given function given by a function pointer function

Parameters

a: double

b: double

bins: a positive integer

function: a pointer to a function

Returns

double