Ginkgo.jl API

Ginkgo.Dense — Type

Dense{T} <: AbstractMatrix{T}

A type for representing dense matrix and vectors. Alias for gko_matrix_dense_eltype_st in C API. where eltype is one of the DataType[Float32, Float64]. For constructing a matrix, it is necessary to provide an executor using the create method.

Examples

# Creating uninitialized vector of length 2, represented as a 2x1 dense matrix
julia> dim = Ginkgo.Dim{2}(2,1); vec1 = Ginkgo.Dense{Float32}(exec, dim)

# Passing a tuple
julia> vec2 = Ginkgo.Dense{Float32}(exec, (2, 1));

# Passing numbers
julia> vec3 = Ginkgo.Dense{Float32}(exec, 2, 1);

# Creating uninitialized dense square matrix
julia> square_mat = Ginkgo.Dense{Float32}(exec, 2);

# Creating initialized dense vector or matrix via reading from a `.mtx` file
julia> b = Ginkgo.Dense{Float32}("b.mtx", exec);

External links

gko::matrix::Dense<T> man page Ginkgo

Ginkgo.Dim — Type

Dim{N}

Abstract type for creating gko::dim<N> like objects

Implementation of operations at the Julia side

External links

gko::dim<N> man page Ginkgo

Ginkgo.Dim2 — Type

Dim2 <: Dim{2}

A type for representing the dimensions of an object. Alias for gko_dim2_st in C API.

Examples

# Creating uninitialized dimensional object
julia> Ginkgo.Dim{2}()
(0, 0)

# Creating initialized dimensional object for identical row and column numbers
julia> Ginkgo.Dim{2}(3)
(3, 3)

# Creaing initialized dimensional object of specified row and column numbers
julia> Ginkgo.Dim{2}(6,9)
(6, 9)

Ginkgo.SparseMatrixCsr — Type

SparseMatrixCsr{Tv, Ti} <: AbstractMatrix{Tv, Ti}

A type for representing sparse matrix and vectors in CSR format. Alias for gko_matrix_csr_eltype_indextype_st in C API. where eltype is one of the DataType[Float32, Float64] and indextype is one of the DataType[Int16, Int32, Int64]. For constructing a matrix, it is necessary to provide an executor using the create method.

Examples

External links

gko::matrix::Csr<ValueType, IndexType> man page Ginkgo

Base.fill! — Method

Base.fill!(mat::Dense{T}, val::G) where {T, G}

Fill the given matrix for all matrix elements with the provided value val

Base.getindex — Method

Base.getindex(mat::Dense{T}, m::Int, n::Int) where T

Obtain an element of the matrix, using Julia indexing

Base.size — Method

Base.size(mat::Dense{T}) where T

Returns the size of the dense matrix/vector as a tuple

Base.size — Method

Base.size(mat::SparseMatrixCsr{Tv,Ti}) where {Tv,Ti}

Returns the size of the sparse matrix/vector as a tuple

Ginkgo.create — Method

create(executor_type::Symbol)

Creation of the executor of a specified executor type.

Parameters

executor_type::Symbol: One of the executor types to create out of supported executor types [:omp, :reference, :cuda]

Ginkgo.elements — Method

elements(mat::Dense{T}) where T

Get number of stored elements of the matrix

Ginkgo.mtx_buffer_str — Method

mtxbufferstr(mat::Dense{T}) where T

Intermediate step that calls gko::write within C level wrapper. Allocates memory temporarily and returns a string pointer in C, then we utilize an IOBuffer to obtain a copy of the allocated cstring in Julia. In the end we deallocate the C string and return the buffered copy.

Ginkgo.nnz — Method

nnz(mat::SparseMatrixCsr{Tv,Ti}) where {Tv,Ti}

Get number of stored elements of the matrix

Ginkgo.norm1! — Method

norm1!(from::Dense{T}, to::Dense{T})

Computes the column-wise Euclidian (L²) norm of this matrix.

External links

void gko::matrix::Dense< ValueType >::compute_norm1 man page Ginkgo

Ginkgo.norm2! — Method

norm2!(from::Dense{T}, to::Dense{T})

Computes the column-wise Euclidian (L²) norm of this matrix.

External links

void gko::matrix::Dense< ValueType >::compute_norm2 man page Ginkgo

Ginkgo.number — Method

number(exec, val::Number)

Initialize a 1x1 matrix representing a number with the provided value val

Ginkgo.spmm! — Method

spmm!(A::SparseMatrixCsr{Tv, Ti}, α::Dense{Tv}, x::Dense{Tv}, β::Dense{Tv}, y::Dense{Tv}) where {Tv, Ti}

x = α*A*b + β*x

Applying to Dense matrices, computes an SpMM product.

Ginkgo.srows — Method

srows(mat::SparseMatrixCsr{Tv,Ti}) where {Tv,Ti}

Returns the number of the srow stored elements (involved warps)

Ginkgo.versioninfo — Method

versioninfo()

Obtain the version information and the supported modules of the underlying Ginkgo library.